+++ CHANGELOG.md
@@ -81,6 +81,7 @@ Grunt & Watch tasks:
 # 2.1.2 (2015-08-20)
 
 **Fixes**
+<<<<<<< 350904500dc0256857920fc88984a84ee9ac44d8
 <<<<<<< 65d0d096a2b4b75b95d6288ba17fb3951389520f
 - [Issue #2558](https://github.com/grafana/grafana/issues/2558). DragDrop: Fix for broken drag drop behavior
 - [Issue #2534](https://github.com/grafana/grafana/issues/2534). Templating: fix for setting template variable value via url and having repeated panels or rows
@@ -92,6 +93,9 @@ panels or rows, fixes #2534
 - [Issue #2534](https://github.com/grafana/grafana/issues/2534). Templating: fix for setting template variable value via url and having repeated panels or rows
 >>>>>>> fix(dragdrop): Fix for broken drag drop behavior, fixes #2558
 =======
+=======
+- [Issue #2568](https://github.com/grafana/grafana/issues/2568). AuthProxy: Fix for server side rendering of panel when using auth proxy
+>>>>>>> fix(auth proxy): Fix for server side rendering of panel when using auth proxy, fixes #2568
 - [Issue #2490](https://github.com/grafana/grafana/issues/2490). Graphite: Dashboard import was broken in 2.1 and 2.1.1, working now
 - [Issue #2565](https://github.com/grafana/grafana/issues/2565). TimePicker: Fix for when you applied custom time range it did not refreh dashboard
 fixes #2565